refactor: AttendanceStatus의 패키지를 constants로 옮겨라
