+++ src/main/java/controller/AttendanceController.java
@@ -4,16 +4,13 @@ import domain.ResponseConverter;
 import domain.UpdatedAttendanceSnapshot;
 import domain.constants.AnswerCommand;
 import domain.Attendance;
-import domain.AttendanceStatus;
+import domain.constants.AttendanceStatus;
 import domain.AttendanceSystem;
 import domain.AttendanceSystemFactory;
 import domain.Crew;
 import domain.constants.ExpulsionStatus;
 import domain.constants.UserCommand;
 import dto.AttendanceResponse;
-import dto.ExpulsionCrewResponse;
-import dto.UpdatedAttendanceSnapshotResponse;
-import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
 import java.util.Map;
