+++ src/test/java/domain/CrewTest.java
@@ -3,6 +3,7 @@ package domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import domain.constants.AttendanceStatus;
 import domain.constants.ExpulsionStatus;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
